pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "pyzbar>=0.1.9",
    "sendgrid>=6.12.5",
    "streamlit>=1.49.1",
    "xlsxwriter>=3.2.9",
]
